Implement exam schedule
![device-2021-09-24-165920.png](https://dev.azure.com/genxas-dev/a8034f85-a478-4e56-823f-80e2d853c459/_apis/git/repositories/db63ddfe-0988-47c7-902c-0d831b9f68c4/pullRequests/10864/attachments/device-2021-09-24-165920.png) ![device-2021-09-24-165928.png](https://dev.azure.com/genxas-dev/a8034f85-a478-4e56-823f-80e2d853c459/_apis/git/repositories/db63ddfe-0988-47c7-902c-0d831b9f68c4/pullRequests/10864/attachments/device-2021-09-24-165928.png) ![device-2021-09-24-170021.png](https://dev.azure.com/genxas-dev/a8034f85-a478-4e56-823f-80e2d853c459/_apis/git/repositories/db63ddfe-0988-47c7-902c-0d831b9f68c4/pullRequests/10864/attachments/device-2021-09-24-170021.png)

+++ library/src/main/java/whiz/sspark/library/data/entity/Term.kt
@@ -6,8 +6,8 @@ import java.util.*
 data class Term(
         @SerializedName("id")val id: String = "",
         @SerializedName("room") val room: Int? = null,
-        @SerializedName("term") val term: Long = 0,
-        @SerializedName("year") val year: Long = 0,
+        @SerializedName("term") val term: Int = 0,
+        @SerializedName("year") val year: Int = 0,
         @SerializedName("startDate") val startDate: Date = Date(),
         @SerializedName("endDate") val endDate: Date = Date(),
         @SerializedName("academicGrade") val academicGrade: Int? = null
